fixed some bugs

+++ data_generator.py
@@ -21,7 +21,7 @@ class DataGenerator:
 
 
         rng = np.random.default_rng(seed)  # seed for generating error positions
-        error_positions = rng.integers(0, len(input_data), size=error_rate, dtype=np.uint8)  # generate positions of bit flips
+        error_positions = rng.integers(0, len(input_data), size=error_rate)  # generate positions of bit flips
         input_data[error_positions] ^= True  # flip bits
